Refactoring to support multiple apps
