+++ src/hardware/pounder/attenuators.rs
