+++ src/hardware/pounder/timestamp.rs
@@ -26,7 +26,7 @@ use stm32h7xx_hal as hal;
 
 use hal::dma::{dma::DmaConfig, PeripheralToMemory, Transfer};
 
-use crate::{timers, SAMPLE_BUFFER_SIZE};
+use crate::{hardware::timers, SAMPLE_BUFFER_SIZE};
 
 // Three buffers are required for double buffered mode - 2 are owned by the DMA stream and 1 is the
 // working data provided to the application. These buffers must exist in a DMA-accessible memory
@@ -89,7 +89,7 @@ impl Timestamper {
         input_capture.listen_dma();
 
         // The data transfer is always a transfer of data from the peripheral to a RAM buffer.
-        let mut data_transfer: Transfer<_, _, PeripheralToMemory, _> =
+        let data_transfer: Transfer<_, _, PeripheralToMemory, _> =
             Transfer::init(
                 stream,
                 input_capture,
@@ -100,8 +100,6 @@ impl Timestamper {
                 config,
             );
 
-        data_transfer.start(|capture_channel| capture_channel.enable());
-
         Self {
             timer: timestamp_timer,
             transfer: data_transfer,
@@ -110,6 +108,12 @@ impl Timestamper {
             // else in the module.
             next_buffer: unsafe { Some(&mut BUF[2]) },
         }
+    }
+
+    /// Start the DMA transfer for collecting timestamps.
+    pub fn start(&mut self) {
+        self.transfer
+            .start(|capture_channel| capture_channel.enable());
     }
 
     /// Update the period of the underlying timestamp timer.
